Various optimisations.
Added an optimised, less accurate Atan2 method
Aggressively inlined various high-use methods.
Changed the ParticleBuffer class to use a pointer copy instead of memcpy
for copying the particles to unmanaged memory. On newer version of .Net,
it compiles down to SSE/AVX instructions, and is the fastest for the amounts
of memory we're dealing with here.
Added a Reset method to the ParticleBuffer class to enable clients to
immediately reuse a buffer.